modularize project structure with llm, tools, and agent
