+++ agent/agent_initializer.py
@@ -1,0 +1,10 @@
+# agent/agent_initializer.py
+from langchain.agents import initialize_agent, AgentType
+
+def create_agent(llm, tools):
+    return initialize_agent(
+        tools=tools,
+        llm=llm,
+        agent_type=AgentType.ZERO_SHOT_REACT_DESCRIPTION,
+        verbose=True
+    )
