+++ llm/ollama_llm.py
@@ -1,0 +1,4 @@
+from langchain_ollama import ChatOllama # Learn about this library later
+
+def get_llm(module_name="qwen2.5:3b"):
+    return ChatOllama(model=module_name)
